requirements.txt
EasyProcess==0.3
entrypoint2==0.2.4
mss==6.1.0
Pillow==8.3.1
pyscreenshot==3.0
pywin32==301
tk==0.1.0
